Suitable
Hi Carlos,

The physics engine in SpriteKit will support your goal. Essentially this project will be an experience with the software development process – using the tools and project management aspects. The code (aside from learning the API) will be relatively straightforward. I'm looking forward to seeing the finished application!

This is a level 4 proposal.

Mr. Gordon